+++ internal/pkg/ociimage/util.go
@@ -6,10 +6,7 @@
 package ociimage
 
 import (
-	"bytes"
-	"context"
 	"fmt"
-	"runtime"
 	"strings"
 
 	"github.com/containers/image/v5/docker"
@@ -19,9 +16,6 @@ import (
 	ocilayout "github.com/containers/image/v5/oci/layout"
 	"github.com/containers/image/v5/signature"
 	"github.com/containers/image/v5/types"
-	ggcrv1 "github.com/google/go-containerregistry/pkg/v1"
-	v1 "github.com/google/go-containerregistry/pkg/v1"
-	"github.com/sylabs/singularity/v4/pkg/sylog"
 )
 
 // defaultPolicy is Singularity's default OCI signature verifiction policy - accept anything.
@@ -59,59 +53,4 @@ func ParseImageRef(imageRef string) (types.ImageReference, error) {
 	}
 
 	return srcRef, nil
-}
-
-// sysCtxToPlatform translates the xxxChoice values in a containers/image
-// types.SytemContext to a go-containerregistry v1.Platform.
-func sysCtxToPlatform(sysCtx *types.SystemContext) ggcrv1.Platform {
-	os := sysCtx.OSChoice
-	if os == "" {
-		os = runtime.GOOS
-	}
-	arch := sysCtx.ArchitectureChoice
-	if arch == "" {
-		arch = runtime.GOARCH
-	}
-	variant := sysCtx.VariantChoice
-	if variant == "" {
-		variant = cpuVariant()
-	}
-	return ggcrv1.Platform{
-		Architecture: arch,
-		Variant:      variant,
-		OS:           os,
-	}
-}
-
-// CheckImageRefPlatform ensures that an image reference satisfies platform requirements in sysCtx
-func CheckImageRefPlatform(ctx context.Context, sysCtx *types.SystemContext, imageRef types.ImageReference) error {
-	if sysCtx == nil {
-		return fmt.Errorf("internal error: sysCtx is nil")
-	}
-	img, err := imageRef.NewImage(ctx, sysCtx)
-	if err != nil {
-		return err
-	}
-	defer img.Close()
-
-	rawConfig, err := img.ConfigBlob(ctx)
-	if err != nil {
-		return err
-	}
-	cf, err := v1.ParseConfigFile(bytes.NewBuffer(rawConfig))
-	if err != nil {
-		return err
-	}
-
-	if cf.Platform() == nil {
-		sylog.Warningf("OCI image doesn't declare a platform. It may not be compatible with this system.")
-		return nil
-	}
-
-	requiredPlatform := sysCtxToPlatform(sysCtx)
-	if cf.Platform().Satisfies(requiredPlatform) {
-		return nil
-	}
-
-	return fmt.Errorf("image (%s) does not satisfy required platform (%s)", cf.Platform(), requiredPlatform)
 }